requirements.txt
Flask>=0.7.2
requests>=0.6.4
calendar
lxml>=2.3.1
BeautifulSoup
pytz
gdata>=2.0
PyYAML==3.10
